requirements.txt
APScheduler==3.6.3
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
dnspython==1.16.0
emoji==1.2.0
flake8==3.9.0
idna==2.10
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
pymongo==3.11.4
python-telegram-bot==13.4.1
pytz==2021.1
requests==2.25.1
six==1.15.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
